requirements.txt
functions-framework==3.*
google-auth==2.23.3
google-cloud-secret-manager==2.19.0
google-api-python-client==2.123.0